+++ Cqrs.Simplified/ICommand.cs
@@ -1,0 +1,4 @@
+﻿namespace Prognetics.CQRS;
+
+public interface ICommand
+{ }
