+++ Cqrs.Simplified/IHandlerResolver.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Prognetics.CQRS;
+
+public interface IHandlerResolver
+{
+    T Resolve<T>();
+    object Resolve(Type t);
+}
