+++ Cqrs.Simplified/IQuery.cs
@@ -1,0 +1,4 @@
+﻿namespace Prognetics.CQRS;
+
+public interface IQuery<T>
+{ }
