+++ Cqrs.Simplified/IQueryHandler.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace Prognetics.CQRS;
+
+public interface IQueryHandler<TQuery, TResult> where TQuery : IQuery<TResult>
+{
+    Task<TResult> Handle(TQuery query);
+}
