+++ Cqrs.Simplified/Prognetics.CQRS.Simplified.csproj
