+++ Prognetics.CQRS.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33122.133
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prognetics.CQRS", "Cqrs\Prognetics.CQRS.csproj", "{20B73295-07A2-459A-9266-27FBB5E74F2A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prognetics.CQRS.Simplified", "Cqrs.Simplified\Prognetics.CQRS.Simplified.csproj", "{3BEBC5D7-A5EA-4838-ABB4-B3E0298E74C9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prognetics.CQRS.Tests", "Tests\Cqrs.Tests\Prognetics.CQRS.Tests.csproj", "{D66E1141-7F0E-4A64-A732-6498E809A3FA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prognetics.CQRS.Simplified.Tests", "Tests\Cqrs.Tests.Simplified\Prognetics.CQRS.Simplified.Tests.csproj", "{2E6EC5CE-C142-466A-8C26-9D1E6231E148}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{20B73295-07A2-459A-9266-27FBB5E74F2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{20B73295-07A2-459A-9266-27FBB5E74F2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{20B73295-07A2-459A-9266-27FBB5E74F2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{20B73295-07A2-459A-9266-27FBB5E74F2A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D66E1141-7F0E-4A64-A732-6498E809A3FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D66E1141-7F0E-4A64-A732-6498E809A3FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D66E1141-7F0E-4A64-A732-6498E809A3FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D66E1141-7F0E-4A64-A732-6498E809A3FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3BEBC5D7-A5EA-4838-ABB4-B3E0298E74C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3BEBC5D7-A5EA-4838-ABB4-B3E0298E74C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3BEBC5D7-A5EA-4838-ABB4-B3E0298E74C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3BEBC5D7-A5EA-4838-ABB4-B3E0298E74C9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E6EC5CE-C142-466A-8C26-9D1E6231E148}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E6EC5CE-C142-466A-8C26-9D1E6231E148}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E6EC5CE-C142-466A-8C26-9D1E6231E148}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E6EC5CE-C142-466A-8C26-9D1E6231E148}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
