+++ Tests/Cqrs.Tests.Simplified/Integration/Command/SumCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Integration.Command
+{
+    public class SumCommand : ICommand
+    {
+        public SumCommand(int number)
+        {
+            Number = number;
+        }
+
+        public int Number { get; }
+    }
+}
