+++ Tests/Cqrs.Tests.Simplified/Integration/Event/EventTests.cs
@@ -1,0 +1,33 @@
+﻿using System.Threading.Tasks;
+using Autofac;
+using Prognetics.CQRS.Simplified.Tests.Shared.Modules;
+using Xunit;
+
+namespace Prognetics.CQRS.Simplified.Tests.Integration.Event
+{
+    public class EventTests
+    {
+        private readonly IContainer _container;
+
+        public EventTests()
+        {
+            _container = BuildContainer();
+        }
+
+        [Fact]
+        public async Task FireEvent()
+        {
+            using var scope = _container.BeginLifetimeScope();
+            var mediator = scope.Resolve<IMediator>();
+
+            await mediator.Publish(new NumberEvent(2));
+        }
+
+        private IContainer BuildContainer()
+        {
+            var containerBuilder = new ContainerBuilder();
+            containerBuilder.RegisterModule(new CqrsModule("Prognetics.CQRS.Simplified.Tests"));
+            return containerBuilder.Build();
+        }
+    }
+}
