+++ Tests/Cqrs.Tests.Simplified/Integration/Event/NumberEvent.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Integration.Event
+{
+    public class NumberEvent : IEvent
+    {
+        public NumberEvent(int number)
+        {
+            Number = number;
+        }
+
+        public int Number { get; }
+    }
+}
