+++ Tests/Cqrs.Tests.Simplified/Integration/Event/SumNumberEventHandler.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+
+namespace Prognetics.CQRS.Simplified.Tests.Integration.Event
+{
+    public class SumNumberEventHandler : IEventHandler<NumberEvent>
+    {
+        public Task Handle(NumberEvent @event)
+        {
+            var result = @event.Number + 3;
+            return Task.CompletedTask;
+        }
+    }
+}
