+++ Tests/Cqrs.Tests.Simplified/Integration/Query/QueryTests.cs
@@ -1,0 +1,51 @@
+﻿using Autofac;
+using FluentAssertions;
+using Prognetics.CQRS.Simplified.Tests.Shared.Modules;
+using Xunit;
+
+namespace Prognetics.CQRS.Simplified.Tests.Integration.Query
+{
+    public class QueryTests
+    {
+        private readonly IContainer _container;
+
+        public QueryTests()
+        {
+            _container = BuildContainer();
+        }
+
+        [Fact]
+        public void ExecuteQueryWithReflection()
+        {
+            using (var scope = _container.BeginLifetimeScope())
+            {
+                var mediator = scope.Resolve<IMediator>();
+
+                var result = mediator.Fetch<SumQuery, int>(new SumQuery(2)).Result;
+
+                result.Should().Be(5);
+            }
+        }
+
+        [Fact]
+        public void ExecuteQueryWithoutReflection()
+        {
+            using (var scope = _container.BeginLifetimeScope())
+            {
+                var mediator = scope.Resolve<IMediator>();
+
+                var query = new SumQuery(2);
+                var result = mediator.Fetch<SumQuery, int>(query).Result;
+
+                result.Should().Be(5);
+            }
+        }
+
+        private IContainer BuildContainer()
+        {
+            var containerBuilder = new ContainerBuilder();
+            containerBuilder.RegisterModule(new CqrsModule("Prognetics.CQRS.Simplified.Tests"));
+            return containerBuilder.Build();
+        }
+    }
+}
