+++ Tests/Cqrs.Tests.Simplified/Integration/Query/SumQueryHandler.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace Prognetics.CQRS.Simplified.Tests.Integration.Query
+{
+    public class SumQueryHandler : IQueryHandler<SumQuery, int>
+    {
+        public Task<int> Handle(SumQuery query)
+        {
+            return Task.FromResult(query.Number + 3);
+        }
+    }
+}
