+++ Tests/Cqrs.Tests.Simplified/Integration/QueryAsync/SumQueryAsync.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Integration.QueryAsync
+{
+    public class SumQueryAsync : IQuery<int>
+    {
+        public SumQueryAsync(int number)
+        {
+            Number = number;
+        }
+
+        public int Number { get; }
+    }
+}
