+++ Tests/Cqrs.Tests.Simplified/Integration/QueryAsync/SumQueryHandlerAsync.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace Prognetics.CQRS.Simplified.Tests.Integration.QueryAsync
+{
+    public class SumQueryHandlerAsync : IQueryHandler<SumQueryAsync, int>
+    {
+        public Task<int> Handle(SumQueryAsync query)
+        {
+            return Task.FromResult(2 + query.Number);
+        }
+    }
+}
