+++ Tests/Cqrs.Tests.Simplified/Shared/GenericCommand/SimpleData.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Shared.GenericCommand
+{
+    public class SimpleData
+    {
+        public SimpleData(string data)
+        {
+            Data = data;
+        }
+
+        public string Data { get; }
+    }
+}
