+++ Tests/Cqrs.Tests.Simplified/Shared/GenericCommand/SomeEntity.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Shared.GenericCommand
+{
+    public class SomeEntity
+    {
+        public SomeEntity(string data)
+        {
+            Data = data;
+        }
+
+        public string Data { get; }
+    }
+}
