+++ Tests/Cqrs.Tests.Simplified/Shared/GenericCommand/TestGenericCommand.cs
@@ -1,0 +1,12 @@
+﻿namespace Prognetics.CQRS.Simplified.Tests.Shared.GenericCommand
+{
+    public class TestGenericCommand<T> : ICommand
+    {
+        public TestGenericCommand(T data)
+        {
+            Data = data;
+        }
+
+        public T Data { get; }
+    }
+}
