--- Tests/Cqrs.Tests/Integration/Command/SumCommand.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Integration.Command
-{
-    public class SumCommand : ICommand
-    {
-        public SumCommand(int number)
-        {
-            Number = number;
-        }
-
-        public int Number { get; }
-    }
-}
