--- Tests/Cqrs.Tests/Integration/Event/NumberEvent.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Integration.Event
-{
-    public class NumberEvent : IEvent
-    {
-        public NumberEvent(int number)
-        {
-            Number = number;
-        }
-
-        public int Number { get; }
-    }
-}
