--- Tests/Cqrs.Tests/Integration/Event/SumNumberEventHandler.cs
@@ -1,13 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Prognetics.CQRS.Tests.Integration.Event
-{
-    public class SumNumberEventHandler : IEventHandler<NumberEvent>
-    {
-        public Task Handle(NumberEvent @event)
-        {
-            var result = @event.Number + 3;
-            return Task.CompletedTask;
-        }
-    }
-}
