--- Tests/Cqrs.Tests/Integration/Query/SumQuery.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Integration.Query
-{
-    public class SumQuery : IQuery<int>
-    {
-        public SumQuery(int number)
-        {
-            Number = number;
-        }
-
-        public int Number { get; }
-    }
-}
