--- Tests/Cqrs.Tests/Integration/QueryAsync/SumQueryHandlerAsync.cs
@@ -1,12 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Prognetics.CQRS.Tests.Integration.QueryAsync
-{
-    public class SumQueryHandlerAsync : IQueryHandler<SumQueryAsync, int>
-    {
-        public Task<int> Handle(SumQueryAsync query)
-        {
-            return Task.FromResult(2 + query.Number);
-        }
-    }
-}
