--- Tests/Cqrs.Tests/Shared/GenericCommand/SimpleData.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Shared.GenericCommand
-{
-    public class SimpleData
-    {
-        public SimpleData(string data)
-        {
-            Data = data;
-        }
-
-        public string Data { get; }
-    }
-}
