--- Tests/Cqrs.Tests/Shared/GenericCommand/SomeEntity.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Shared.GenericCommand
-{
-    public class SomeEntity
-    {
-        public SomeEntity(string data)
-        {
-            Data = data;
-        }
-
-        public string Data { get; }
-    }
-}
