--- Tests/Cqrs.Tests/Shared/GenericCommand/TestGenericCommandHandler.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace Prognetics.CQRS.Tests.Shared.GenericCommand
-{
-    public class TestGenericCommandHandler<T> : ICommandHandler<TestGenericCommand<T>>
-    {
-        public Task Handle(TestGenericCommand<T> command)
-        {
-            Console.WriteLine(command.Data);
-            return Task.CompletedTask;
-        }
-    }
-}
