--- Tests/Cqrs.Tests/Shared/GenericQuery/SampleGenericQuery.cs
@@ -1,12 +1,0 @@
-﻿namespace Prognetics.CQRS.Tests.Shared.GenericQuery
-{
-    public class SampleGenericQuery<T> : IQuery<int>
-    {
-        public SampleGenericQuery(T data)
-        {
-            Data = data;
-        }
-
-        public T Data { get; }
-    }
-}
